authed routes

+++ src/app/api/task/route.ts
@@ -1,10 +1,14 @@
 import { NextResponse } from "next/server";
-import { authMW } from "../auth";
+import { authMW, INTERNAL_UID_HEADER } from "../auth";
 import { pipe } from "../nextApi";
 import taskrService from "../taskr-service";
 
-export const POST = pipe(authMW)(async function (req) {
-  const { id, task } = await req.json();
+export const POST = pipe(authMW)(async (req) => {
+  const id = req.headers.get(INTERNAL_UID_HEADER);
+  if (!id) {
+    return new NextResponse(null, { status: 404 });
+  }
+  const { task } = await req.json();
 
   if (!id || !task) {
     return new NextResponse(null, { status: 400 });
@@ -23,34 +27,44 @@ export const POST = pipe(authMW)(async function (req) {
   });
 });
 
-export async function PUT(req: Request) {
-  const { id, task } = await req.json();
+export const PUT = pipe(authMW)(async (req) => {
+  const id = req.headers.get(INTERNAL_UID_HEADER);
+  if (!id) {
+    return new NextResponse(null, { status: 404 });
+  }
+
+  const { task } = await req.json();
 
   if (!id || !task) {
-    return new Response(null, { status: 400 });
+    return new NextResponse(null, { status: 400 });
   }
 
   const res = await taskrService.updateTasks(id, task);
 
   if (res.ok()) {
-    return Response.json(res.data, {
+    return NextResponse.json(res.data, {
       status: 200,
     });
   }
 
-  return new Response(null, {
+  return new NextResponse(null, {
     status: res.code,
   });
-}
+});
 
-export async function DELETE(req: Request) {
-  const { id, taskId } = await req.json();
+export const DELETE = pipe(authMW)(async (req) => {
+  const id = req.headers.get(INTERNAL_UID_HEADER);
+  if (!id) {
+    return new NextResponse(null, { status: 404 });
+  }
+  
+  const { taskId } = await req.json();
 
   if (!id || !taskId) {
-    return new Response(null, { status: 400 });
+    return new NextResponse(null, { status: 400 });
   }
 
   const res = await taskrService.deleteTask(id, taskId);
 
-  return new Response(null, { status: res.code });
-}
+  return new NextResponse(null, { status: res.code });
+});
